これらもstateの操作に関連したものなので移動
