+++ src/mandelbrot.ts
@@ -61,14 +61,6 @@ export const calcVars = (
   };
 };
 
-export const cloneParams = (params: MandelbrotParams): MandelbrotParams => ({
-  x: BigNumber(params.x),
-  y: BigNumber(params.y),
-  r: BigNumber(params.r),
-  N: params.N,
-  mode: params.mode,
-});
-
 export const startCalculation = async (
   onComplete: (elapsed: number) => void,
   onTranslated: () => void,
@@ -157,6 +149,3 @@ export const startCalculation = async (
     shouldReuseRefOrbit: getStore("shouldReuseRefOrbit"),
   });
 };
-
-export const isSameParams = (a: MandelbrotParams, b: MandelbrotParams) =>
-  a.x === b.x && a.y === b.y && a.r === b.r && a.N === b.N && a.mode === b.mode;
