+++ src/view/right-sidebar/poi.tsx
@@ -1,10 +1,12 @@
 import { Button } from "@/components/ui/button";
 import { ScrollArea } from "@/components/ui/scroll-area";
-import { getCurrentParams } from "@/mandelbrot-state/mandelbrot-state";
+import {
+  cloneParams,
+  getCurrentParams,
+} from "@/mandelbrot-state/mandelbrot-state";
 import { IconCirclePlus } from "@tabler/icons-react";
 import throttle from "lodash.throttle";
 import { useEffect, useRef } from "react";
-import { cloneParams } from "../../mandelbrot";
 import { POICard } from "./poi-card";
 import { usePOI } from "./use-poi";
